# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-floatmultiply",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — compute `a * b`.",
        "description": "Input validation is delegated to `srvcs-isnumber` over HTTP (the single\nsource of truth for \"is this a number\"), once per operand. Both integer and\nfractional inputs are accepted and coerced with `as_f64`; the result is an\n`f64`. If the dependency is unreachable, this service reports itself degraded\nrather than guessing.",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "422": {
            "description": "an operand is not a number"
          },
          "500": {
            "description": "an operand passed validation but is not representable as f64"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EvalRequest": {
        "type": "object",
        "required": [
          "a",
          "b"
        ],
        "properties": {
          "a": {
            "type": "object"
          },
          "b": {
            "type": "object"
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      },
      "ProductResponse": {
        "type": "object",
        "required": [
          "a",
          "b",
          "result"
        ],
        "properties": {
          "a": {
            "type": "object"
          },
          "b": {
            "type": "object"
          },
          "result": {
            "type": "number",
            "format": "double",
            "description": "The product `a * b` as a floating-point number."
          }
        }
      }
    }
  }
}